Consolidate JsonExtractStaticTests into JsonExtractTests
Move all static edge-case tests into JsonExtractTests as inline
testXxx() methods, eliminating duplicated infrastructure borrowed
from AbstractScalarFunctionTestCase. This follows the SubstringTests
pattern of mixing parameterized suppliers with inline test methods.
